requirements.txt
huepy==0.9.8.1
PyQRCode==1.2.1
pypng==0.0.18
